refactor redux directory
