+++ src/components/forms/sections/ApplicationSelector/ApplicationSelector.tsx
@@ -2,7 +2,7 @@ import React, { FC, useState } from 'react';
 import { connect } from 'react-redux';
 import { Label } from 'reactstrap';
 import { SELECTED_BERTH_LIMIT } from '../../../../constants/berths';
-import { Store } from '../../../../types/ducks';
+import { Store } from '../../../../redux/types';
 import Alert from '../../../common/Alert';
 import Input from '../../../common/Input';
