--- src/ducks/index.ts
@@ -1,2 +1,0 @@
-export { default as forms } from './forms';
-export { default as berths } from './berths';
