+++ src/redux/actions/FormActions.tsx
@@ -1,0 +1,6 @@
+import { createAction } from 'redux-actions';
+
+const onSubmit = createAction('SUBMIT_FORM', (formData: {}) => formData);
+const resetValues = createAction('RESET_FORM');
+
+export { onSubmit, resetValues };
