+++ src/redux/actions/index.tsx
@@ -1,0 +1,7 @@
+import * as BerthActions from './BerthActions';
+import * as FormActions from './FormActions';
+
+export default {
+  BerthActions,
+  FormActions
+};
