+++ src/redux/reducers/index.tsx
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import BerthReducers from './BerthReducers';
+import FormReducers from './FormReducers';
+
+export default combineReducers({
+  berths: BerthReducers,
+  forms: FormReducers
+});
